V1: Add conversion for TaskRef.Bundle
This commit adds support for TaskRef.Bundle when converting between
v1beta1 and v1 versions of TaskRef. This allows us to release v1 TaskRun
and PipelineRun in a backwards compatible way by ensuring that v1beta1
TaskRuns and PipelineRuns with Bundle serialized into annotations on the
v1 TaskRuns and PipelineRuns on conversion.